+++ rio_tiler/io/xarray.py
@@ -9,7 +9,6 @@ import attr
 from morecantile import Tile, TileMatrixSet
 from rasterio.crs import CRS
 from rasterio.enums import Resampling
-from rasterio.rio.overview import get_maximum_overview_level
 from rasterio.transform import from_bounds, rowcol
 from rasterio.warp import calculate_default_transform
 from rasterio.warp import transform as transform_coords
@@ -65,9 +64,6 @@ class XarrayReader(BaseReader):
     tms: TileMatrixSet = attr.ib(default=WEB_MERCATOR_TMS)
     geographic_crs: CRS = attr.ib(default=WGS84_CRS)
 
-    _minzoom: int = attr.ib(init=False, default=None)
-    _maxzoom: int = attr.ib(init=False, default=None)
-
     _dims: List = attr.ib(init=False, factory=list)
 
     def __attrs_post_init__(self):
@@ -93,91 +89,25 @@ class XarrayReader(BaseReader):
                 f"Invalid geographic bounds: {self.bounds}. Must be within (-180, -90, 180, 90)."
             )
 
+        self.transform = self.input.rio.transform()
+        self.height = self.input.rio.height
+        self.width = self.input.rio.width
+
         self._dims = [
             d
             for d in self.input.dims
             if d not in [self.input.rio.x_dim, self.input.rio.y_dim]
         ]
 
-    def _dst_geom_in_tms_crs(self):
-        """Return dataset info in TMS projection."""
-        tms_crs = self.tms.rasterio_crs
-        if self.crs != tms_crs:
-            dst_affine, w, h = calculate_default_transform(
-                self.crs,
-                tms_crs,
-                self.input.rio.width,
-                self.input.rio.height,
-                *self.bounds,
-            )
-        else:
-            dst_affine = list(self.input.rio.transform())
-            w = self.input.rio.width
-            h = self.input.rio.height
-
-        return dst_affine, w, h
-
-    def get_minzoom(self) -> int:
-        """Define dataset minimum zoom level."""
-        if self._minzoom is None:
-            # We assume the TMS tilesize to be constant over all matrices
-            # ref: https://github.com/OSGeo/gdal/blob/dc38aa64d779ecc45e3cd15b1817b83216cf96b8/gdal/frmts/gtiff/cogdriver.cpp#L274
-            tilesize = self.tms.tileMatrices[0].tileWidth
-
-            try:
-                dst_affine, w, h = self._dst_geom_in_tms_crs()
-
-                # The minzoom is defined by the resolution of the maximum theoretical overview level
-                # We assume `tilesize`` is the smallest overview size
-                overview_level = get_maximum_overview_level(w, h, minsize=tilesize)
-
-                # Get the resolution of the overview
-                resolution = max(abs(dst_affine[0]), abs(dst_affine[4]))
-                ovr_resolution = resolution * (2**overview_level)
-
-                # Find what TMS matrix match the overview resolution
-                self._minzoom = self.tms.zoom_for_res(ovr_resolution)
-
-            except:  # noqa
-                # if we can't get min/max zoom from the dataset we default to TMS maxzoom
-                warnings.warn(
-                    "Cannot determine maxzoom based on dataset information, will default to TMS maxzoom.",
-                    UserWarning,
-                )
-                self._minzoom = self.tms.maxzoom
-
-        return self._minzoom
-
-    def get_maxzoom(self) -> int:
-        """Define dataset maximum zoom level."""
-        if self._maxzoom is None:
-            try:
-                dst_affine, _, _ = self._dst_geom_in_tms_crs()
-
-                # The maxzoom is defined by finding the minimum difference between
-                # the raster resolution and the zoom level resolution
-                resolution = max(abs(dst_affine[0]), abs(dst_affine[4]))
-                self._maxzoom = self.tms.zoom_for_res(resolution)
-
-            except:  # noqa
-                # if we can't get min/max zoom from the dataset we default to TMS maxzoom
-                warnings.warn(
-                    "Cannot determine maxzoom based on dataset information, will default to TMS maxzoom.",
-                    UserWarning,
-                )
-                self._maxzoom = self.tms.maxzoom
-
-        return self._maxzoom
-
     @property
     def minzoom(self):
         """Return dataset minzoom."""
-        return self.get_minzoom()
+        return self._minzoom
 
     @property
     def maxzoom(self):
         """Return dataset maxzoom."""
-        return self.get_maxzoom()
+        return self._maxzoom
 
     @property
     def band_names(self) -> List[str]: